ScheduledStrategy.py part 1 before testing
