+++ athena-scripts/raw_stage_optimisation_solution/scripts/etl/strategies/BackfillStrategy.py
@@ -1,0 +1,12 @@
+from .Strategy import Strategy
+
+
+class BackfillStrategy(Strategy):
+    def extract(self):
+        pass
+
+    def transform(self, df):
+        pass
+
+    def load(self):
+        pass
